refactor: 파일명 변경 & 디렉토리 변경
- AuthController -> AuthTestController
- api -> api/test
